+++ include/ftl/function.h
@@ -30,369 +30,46 @@
 #ifndef FTL_FUNCTION_H
 #define FTL_FUNCTION_H
 
-#include <memory>
-#include <stdexcept>
-#include <functional>
-#include "type_functions.h"
-
-#ifdef __GNUC__
-#pragma GCC diagnostic push
-#pragma GCC diagnostic ignored "-Wstrict-aliasing"
-#endif
+#include "implementation/function.h"
 
 namespace ftl {
 
-	template<typename> class function;
-
-	template<typename>
-	struct force_function_heap_allocation : std::false_type {};
-
-	// Namespace for internal details
-	namespace _dtl {
-		struct manager_storage_type;
-
-		struct functor_padding {
-		protected:
-			size_t padding_fst;
-			size_t padding_snd;
-		};
-
-		struct empty_struct {};
-
-		template<typename R, typename...Ps>
-		R empty_call(const functor_padding&, Ps...) {
-			throw std::bad_function_call();
-		}
-
-		template<typename T, typename Alloc>
-		struct is_inplace_allocated {
-			static constexpr bool value
-				// so that it fits
-				= sizeof(T) <= sizeof(functor_padding)
-
-				// so that it will be aligned
-				&& std::alignment_of<functor_padding>::value % std::alignment_of<T>::value == 0
-
-				// so that we can offer noexcept move
-				&& std::is_nothrow_move_constructible<T>::value
-
-				// so that the user can override it
-				&& !force_function_heap_allocation<T>::value;
-		};
-
-		template<typename T>
-		T to_functor(T&& func) {
-			return std::forward<T>(func);
-		}
-		template<typename R, typename C, typename...Ps>
-		auto to_functor(R (C::*func)(Ps...)) -> decltype(std::mem_fn(func)) {
-			return std::mem_fn(func);
-		}
-		template<typename R, typename C, typename...Ps>
-		auto to_functor(R (C::*func)(Ps...) const) -> decltype(std::mem_fn(func)) {
-			return std::mem_fn(func);
-		}
-
-		template<typename T>
-		struct functor_type {
-			using type = decltype(to_functor(std::declval<T>()));
-		};
-
-		template<typename T>
-		bool is_null(const T&) {
-			return false;
-		}
-
-		template<typename R, typename...Ps>
-		bool is_null(R (* const & fp)(Ps...)) {
-			return fp == nullptr;
-		}
-
-		template<typename R, typename C, typename...Ps>
-		bool is_null(R (C::* const & fp)(Ps...)) {
-			return fp == nullptr;
-		}
-
-		template<typename R, typename C, typename...Ps>
-		bool is_null(R (C::* const & fp)(Ps...) const) {
-			return fp == nullptr;
-		}
-
-		template<typename, typename>
-		struct is_valid_function_argument {
-			static constexpr bool value = false;
-		};
-
-		template<typename R, typename...Ps>
-		struct is_valid_function_argument<function<R(Ps...)>, R (Ps...)> {
-			static constexpr bool value = false;
-		};
-
-		template<typename T, typename R, typename...Ps>
-		struct is_valid_function_argument<T, R (Ps...)> {
-
-			template<typename U>
-			static decltype(to_functor(std::declval<U>())(std::declval<Ps>()...)) check(U *);
-
-			template<typename>
-			static empty_struct check(...);
-
-			static constexpr bool value = std::is_convertible<decltype(check<T>(nullptr)), R>::value;
-		};
-
-		enum function_manager_calls
-		{
-			call_move_and_destroy,
-			call_copy,
-			call_copy_functor_only,
-			call_destroy,
-		};
-
-		template<typename T, typename Allocator>
-		void* function_manager(
-				void* first_arg,
-				void* second_arg,
-				function_manager_calls call_type);
-
-		typedef void *(*manager_type)(void*, void*, function_manager_calls);
-
-		struct manager_storage_type {
-
-			template<typename Allocator>
-			Allocator& get_allocator() noexcept {
-				return reinterpret_cast<Allocator&>(manager);
-			}
-
-			template<typename Allocator>
-			const Allocator& get_allocator() const noexcept {
-				return reinterpret_cast<const Allocator&>(manager);
-			}
-
-			functor_padding functor;
-			manager_type manager;
-		};
-
-		template<typename T, typename Allocator, typename Enable = void>
-		struct function_manager_inplace_specialisation {
-
-			template<typename R, typename...Ps>
-			static R call(const functor_padding& storage, Ps... ps) {
-				// do not call get_functor_ref because I want this function to
-				// be fast in debug when nothing gets inlined
-				return
-					reinterpret_cast<T&>(
-						const_cast<functor_padding&>(storage))(std::forward<Ps>(ps)...);
-			}
-
-			static void store_functor(manager_storage_type& storage, T to_store) {
-				new (&get_functor_ref(storage)) T(std::forward<T>(to_store));
-			}
-
-			static void move_functor(manager_storage_type& lhs, manager_storage_type&& rhs) noexcept {
-				new (&get_functor_ref(lhs)) T(std::move(get_functor_ref(rhs)));
-			}
-
-			static void destroy_functor(Allocator &, manager_storage_type & storage) noexcept {
-				get_functor_ref(storage).~T();
-			}
-
-			static T& get_functor_ref(const manager_storage_type& storage) noexcept {
-				return reinterpret_cast<T&>(const_cast<functor_padding&>(storage.functor));
-			}
-		};
-
-		template<typename T, typename Allocator>
-		struct function_manager_inplace_specialisation<
-					T,
-					Allocator,
-					typename std::enable_if<
-						!is_inplace_allocated<T, Allocator>::value>::type> {
-
-			using alloc_traits = std::allocator_traits<Allocator>;
-			using ptr_t = typename alloc_traits::pointer;
-
-			template<typename R, typename...Ps>
-			static R call(const functor_padding& storage, Ps... ps) {
-				// do not call get_functor_ptr_ref because I want this function
-				// to be fast in debug when nothing gets inlined
-				return
-					(*reinterpret_cast<ptr_t&>
-						(const_cast<functor_padding&>(storage)))(std::forward<Ps>(ps)...);
-			}
-
-			static void store_functor(manager_storage_type& self, T to_store) {
-
-				Allocator & allocator = self.get_allocator<Allocator>();;
-				static_assert(
-						sizeof(ptr_t) <= sizeof(self.functor),
-						"The allocator's pointer type is too big");
-
-				ptr_t* ptr = new (&get_functor_ptr_ref(self)) ptr_t(alloc_traits::allocate(allocator, 1));
-
-				alloc_traits::construct(allocator, *ptr, std::forward<T>(to_store));
-			}
-
-			static void move_functor(manager_storage_type& lhs, manager_storage_type&& rhs) noexcept {
-
-				static_assert(
-						std::is_nothrow_move_constructible<ptr_t>::value,
-						"we can't offer a noexcept swap if the pointer type is not nothrow move constructible");
-
-				new (&get_functor_ptr_ref(lhs)) ptr_t(std::move(get_functor_ptr_ref(rhs)));
-
-				// this next assignment makes the destroy function easier
-				get_functor_ptr_ref(rhs) = nullptr;
-			}
-
-			static void destroy_functor(Allocator& allocator, manager_storage_type& storage) noexcept {
-
-				ptr_t& pointer = get_functor_ptr_ref(storage);
-				if (!pointer)
-					return;
-
-				alloc_traits::destroy(allocator, pointer);
-				alloc_traits::deallocate(allocator, pointer, 1);
-			}
-
-			static T & get_functor_ref(const manager_storage_type& storage) noexcept {
-				return *get_functor_ptr_ref(storage);
-			}
-
-			static ptr_t& get_functor_ptr_ref(const manager_storage_type& storage) noexcept {
-				return reinterpret_cast<ptr_t&>(const_cast<functor_padding&>(storage.functor));
-			}
-		};
-
-		template<typename T, typename Allocator>
-		static void create_manager(manager_storage_type& storage, Allocator&& allocator) {
-
-			new (&storage.get_allocator<Allocator>()) Allocator(std::move(allocator));
-			storage.manager = &function_manager<T, Allocator>;
-		}
-
-		// this function acts as a vtable. it is an optimization to prevent
-		// code-bloat from rtti. see the documentation of boost::function
-		template<typename T, typename Allocator>
-		void* function_manager(void* first_arg, void* second_arg, function_manager_calls call_type) {
-
-			using specialisation = function_manager_inplace_specialisation<T, Allocator>;
-
-			static_assert(
-					std::is_empty<Allocator>::value,
-					"the allocator must be an empty class because I don't have space for it");
-
-			switch(call_type) {
-			case call_move_and_destroy: {
-				manager_storage_type& lhs = *static_cast<manager_storage_type*>(first_arg);
-				manager_storage_type& rhs = *static_cast<manager_storage_type*>(second_arg);
-
-				specialisation::move_functor(lhs, std::move(rhs));
-				specialisation::destroy_functor(rhs.get_allocator<Allocator>(), rhs);
-
-				create_manager<T, Allocator>(lhs, std::move(rhs.get_allocator<Allocator>()));
-				rhs.get_allocator<Allocator>().~Allocator();
-
-				return nullptr;
-			}
-			case call_copy: {
-				manager_storage_type& lhs = *static_cast<manager_storage_type*>(first_arg);
-
-				const manager_storage_type & rhs = *static_cast<const manager_storage_type*>(second_arg);
-
-				create_manager<T, Allocator>(lhs, Allocator(rhs.get_allocator<Allocator>()));
-
-				specialisation::store_functor(lhs, const_cast<const T&>(specialisation::get_functor_ref(rhs)));
-
-				return nullptr;
-			}
-			case call_destroy: {
-				manager_storage_type& self = *static_cast<manager_storage_type *>(first_arg);
-
-				specialisation::destroy_functor(self.get_allocator<Allocator>(), self);
-
-				self.get_allocator<Allocator>().~Allocator();
-
-				return nullptr;
-			}
-			case call_copy_functor_only:
-				specialisation::store_functor(
-						*static_cast<manager_storage_type *>(first_arg),
-						const_cast<const T&>(
-							specialisation::get_functor_ref(
-								*static_cast<const manager_storage_type*>(second_arg))));
-
-				return nullptr;
-
-			default:
-				return nullptr;
-			}
-		}
-
-		/* TODO: Make currying work even when we give N > 1, N < Nparams
-		 * arguments to a function of Nparams parameters.
-		 */
-		template<typename...>
-		struct curried {};
-
-		template<typename R>
-		struct curried<R> {
-			R operator()() const {
-				throw(std::logic_error("Curried calling of parameterless function"));
-			}
-		};
-
-		template<typename R, typename P>
-		struct curried<R,P> {
-			function<R()> operator()(P) const {
-				throw(std::logic_error("Curried calling of parameterless function"));
-			}
-		};
-
-		template<typename R, typename P1, typename P2, typename...Ps>
-		struct curried<R,P1,P2,Ps...> {
-        private:
-            /// Apply one argument.
-            using applied_type = function<R(P2,Ps...)>;
-            applied_type apply_one(P1 p1) const {
-				auto self = *reinterpret_cast<const function<R(P1,P2,Ps...)>*>(this);
-				return [self,p1] (P2 p2, Ps...ps) {
-					return self.operator()(p1, std::forward<P2>(p2), std::forward<Ps>(ps)...);
-				};
-			}
-		public:
-			function<R(P2,Ps...)> operator() (P1 p1) const {
-				return apply_one(std::move(p1));
-			}
-
-			// Apply each argument, return a new function.
-			// If the number of arguments equals the function's arity,
-			// ftl::function::operator() will be called instead.
-			template<typename...Ps2>
-			auto operator()(P1 p1, P2 p2, Ps2...ps2) const
-				-> typename std::result_of<applied_type(P2,Ps2...)>::type
-			{
-				applied_type g = apply_one(std::move(p1));
-				return g(std::move(p2),std::move(ps2)...);
-			}
-		};
-
-	}
+	/**
+	 * \defgroup function Function
+	 *
+	 * Generalised function type similar to `std::function`.
+	 *
+	 * \code
+	 *   #include <ftl/function.h>
+	 * \endcode
+	 *
+	 * \par Dependencies
+	 * - `<memory>`
+	 * - `<stdexcept>`
+	 * - `<functional>`
+	 * - \ref typelevel
+	 */
 
 	/**
-	 * Function object encapsulator.
+	 * Generic, type erasing function object encapsulator.
 	 *
 	 * \tparam R Return value of the wrapped function or function object.
 	 * \tparam Ps Parameter pack of the wrapped function's `operator()`.
+	 *
+	 * The point of including this data type is that, unlike `std::function`,
+	 * it provides built-in support for curried calling.
 	 *
 	 * \par Concepts
 	 * - \ref fullycons
 	 * - \ref assignable
 	 * - \ref fn`<R(Ps...)>`
 	 * - \ref fn`<`\ref fn`<R(P2, ..., PN)>(P1)>`, assuming `sizeof...(Ps) > 2`
-	 *   and `P1`, `P2`, ..., `PN` are the elements of `Ps`.
-	 * - \ref applicative
-	 * - \ref functor
-	 * - \ref monoid, if and only if R is a Monoid.
+	 *   and `P1`, `P2`, ..., `PN` are the elements of `Ps`
+	 * - \ref functorpg (instance defined in \ref functional)
+	 * - \ref applicativepg (instance defined in \ref functional)
+	 * - \ref monadpg (instance defined in \ref functional)
+	 * - \ref monoidpg, if and only if `R` is a monoid (instance defined in
+	 *   \ref functional)
 	 *
 	 * If the second \ref fn instance above is not completely obvious in what it
 	 * means, it signifies that ftl::function supports the notion of curried
@@ -405,21 +82,21 @@ namespace ftl {
 	 * \endcode
 	 *
 	 * \note While often convenient, it may give rise to ambiguous calls to
-	 * `operator()` if the function object encapsulated by the ftl::function
-	 * has several overloaded `operator()` that differ only in the number of
+	 * `operator()` if the function object encapsulated by the `ftl::function`
+	 * has several overloaded call operators that differ only in the number of
 	 * parameters they take.
 	 *
 	 * \warning Curried calling _will_ result in copies of the parameter being
 	 *          made. Every time you invoke `operator()` without filling the
 	 *          complete parameter list, you are creating copies.
 	 *
-	 * \ingroup functional
+	 * \ingroup function
 	 */
 	template<typename>
 	class function {};
 
 	template<typename R, typename...Ps>
-	class function<R(Ps...)> : private _dtl::curried<R,Ps...> {
+	class function<R(Ps...)> : private ::ftl::_dtl::curried<R,Ps...> {
 	public:
 		/**
 		 * Type sequence representation of the function's parameter list.
@@ -445,8 +122,8 @@ namespace ftl {
 		function(const function& f) : call(f.call) {
 			f.manager_storage.manager(
 					&manager_storage,
-					const_cast<_dtl::manager_storage_type*>(&f.manager_storage),
-					_dtl::call_copy);
+					const_cast<::ftl::_dtl::manager_storage_type*>(&f.manager_storage),
+					::ftl::_dtl::call_copy);
 		}
 
 		function(function&& f) noexcept {
@@ -454,95 +131,129 @@ namespace ftl {
 			swap(f);
 		}
 
+		/**
+		 * Construct from arbitrary function object.
+		 *
+		 * \tparam F must have a function call operator matching the type the
+		 *           `ftl::function` is declared as.
+		 */
 		template<typename F>
 		function(
 				F f,
 				typename std::enable_if<
-					_dtl::is_valid_function_argument<F, R (Ps...)>::value,
-					_dtl::empty_struct>::type = _dtl::empty_struct())
-		noexcept(_dtl::is_inplace_allocated<
+					::ftl::_dtl::is_valid_function_argument<F, R (Ps...)>::value,
+					::ftl::_dtl::empty_struct>::type = ::ftl::_dtl::empty_struct()
+		)
+		noexcept(::ftl::_dtl::is_inplace_allocated<
 				F,
-				std::allocator<typename _dtl::functor_type<F>::type>>::value) {
-
-			if(_dtl::is_null(f))
+				std::allocator<typename ::ftl::_dtl::functor_type<F>::type>>::value
+		)
+		{
+			if(::ftl::_dtl::is_null(f))
 				initialise_empty();
 
 			else {
-				using functor_type = typename _dtl::functor_type<F>::type;
-				initialise(_dtl::to_functor(std::forward<F>(f)), std::allocator<functor_type>());
+				using functor_type = typename ::ftl::_dtl::functor_type<F>::type;
+				initialise(
+					::ftl::_dtl::to_functor(
+						std::forward<F>(f)
+					),
+					std::allocator<functor_type>()
+				);
 			}
 		}
 
+		/// Default construct using a custom allocator
 		template<typename Allocator>
 		function(std::allocator_arg_t, const Allocator&) {
 			// ignore the allocator because I don't allocate
 			initialise_empty();
 		}
 
+		/// Null construct using a custom allocator
 		template<typename Allocator>
 		function(std::allocator_arg_t, const Allocator&, std::nullptr_t) {
 			// ignore the allocator because I don't allocate
 			initialise_empty();
 		}
 
+		/**
+		 * Construct from arbitrary function object using a custom allocator.
+		 */
 		template<typename Allocator, typename F>
 		function(
 				std::allocator_arg_t,
 				const Allocator& allocator,
 				F functor,
 				typename std::enable_if<
-					_dtl::is_valid_function_argument<F, R (Ps...)>::value,
-					_dtl::empty_struct>::type = _dtl::empty_struct())
-		noexcept(_dtl::is_inplace_allocated<F, Allocator>::value) {
+					::ftl::_dtl::is_valid_function_argument<F, R (Ps...)>::value,
+					::ftl::_dtl::empty_struct>::type = ::ftl::_dtl::empty_struct())
+		noexcept(::ftl::_dtl::is_inplace_allocated<F, Allocator>::value) {
 
-			if(_dtl::is_null(functor))
+			if(::ftl::_dtl::is_null(functor))
 				initialise_empty();
 
 			else {
-				initialise(_dtl::to_functor(std::forward<F>(functor)), Allocator(allocator));
+				initialise(::ftl::_dtl::to_functor(
+					std::forward<F>(functor)), Allocator(allocator)
+				);
 			}
 		}
 
+		/// Copy construct using a custom allocator
 		template<typename Allocator>
-		function(
-				std::allocator_arg_t,
+		function(std::allocator_arg_t,
 				const Allocator& allocator,
-				const function& other)
+				const function& other
+		)
 		: call(other.call) {
 
 			using alloc_traits = std::allocator_traits<Allocator>;
-			using MyAllocator = typename alloc_traits::template rebind_alloc<function>::other;
+			using MyAllocator =
+				typename alloc_traits::template rebind_alloc<function>::other;
 
 			// first try to see if the allocator matches the target type
-			_dtl::manager_type manager_for_allocator
-				= &_dtl::function_manager<
-					typename alloc_traits::value_type, Allocator>;
+			::ftl::_dtl::manager_type manager_for_allocator =
+				&::ftl::_dtl::function_manager<
+					typename alloc_traits::value_type, Allocator
+				>;
 
 			if(other.manager_storage.manager == manager_for_allocator) {
-				_dtl::create_manager<typename alloc_traits::value_type, Allocator>(
-						manager_storage,
-						Allocator(allocator));
+
+				::ftl::_dtl::create_manager<
+					typename alloc_traits::value_type, Allocator
+				> (
+					manager_storage, Allocator(allocator)
+				);
 
 				manager_for_allocator(
-						&manager_storage,
-						const_cast<_dtl::manager_storage_type*>(&other.manager_storage),
-						_dtl::call_copy_functor_only);
+					&manager_storage,
+					const_cast<::ftl::_dtl::manager_storage_type*>(
+						&other.manager_storage
+					),
+					::ftl::_dtl::call_copy_functor_only
+				);
 			}
 
 			// if it does not, try to see if the target contains my type. this
 			// breaks the recursion of the last case. otherwise repeated copies
 			// would allocate more and more memory
 			else if(other.manager_storage.manager
-					== &_dtl::function_manager<function, MyAllocator>) {
+					== &::ftl::_dtl::function_manager<function, MyAllocator>
+			) {
 
-				_dtl::create_manager<function, MyAllocator>(
-						manager_storage,
-						MyAllocator(allocator));
+				::ftl::_dtl::create_manager<function, MyAllocator>(
+					manager_storage,
+					MyAllocator(allocator)
+				);
 
-				_dtl::function_manager<function, MyAllocator>(
-						&manager_storage,
-						const_cast<_dtl::manager_storage_type*>(&other.manager_storage),
-						_dtl::call_copy_functor_only);
+				::ftl::_dtl::function_manager<function, MyAllocator>(
+					&manager_storage,
+					const_cast<::ftl::_dtl::manager_storage_type*>(
+						&other.manager_storage
+					),
+					::ftl::_dtl::call_copy_functor_only
+				);
 			}
 
 			else
@@ -552,9 +263,10 @@ namespace ftl {
 			}
 		}
 
+		/// Move construct using a custom allocator
 		template<typename Allocator>
 		function(std::allocator_arg_t, const Allocator&, function&& other) noexcept {
-			// ignore the allocator because I don't allocate
+			// ignore the allocator because there's no allocation
 			initialise_empty();
 			swap(other);
 		}
@@ -563,7 +275,7 @@ namespace ftl {
 			manager_storage.manager(
 					&manager_storage,
 					nullptr,
-					_dtl::call_destroy);
+					::ftl::_dtl::call_destroy);
 		}
 
 		function& operator= (function other) noexcept {
@@ -571,7 +283,8 @@ namespace ftl {
 			return *this;
 		}
 
-		using _dtl::curried<R,Ps...>::operator();
+		// Inherit the curried function call operator(s)
+		using ::ftl::_dtl::curried<R,Ps...>::operator();
 
 		/// Call the wrapped function object
 		R operator()(Ps...ps) const {
@@ -580,51 +293,51 @@ namespace ftl {
 
 		template<typename F, typename Allocator>
 		void assign(F&& f, const Allocator& alloc)
-		noexcept(_dtl::is_inplace_allocated<F, Allocator>::value) {
+		noexcept(::ftl::_dtl::is_inplace_allocated<F, Allocator>::value) {
 			function(std::allocator_arg, alloc, f).swap(*this);
 		}
 
 		void swap(function& other) noexcept {
-			_dtl::manager_storage_type temp_storage;
+			::ftl::_dtl::manager_storage_type temp_storage;
 
 			other.manager_storage.manager(
 					&temp_storage,
 					&other.manager_storage,
-					_dtl::call_move_and_destroy);
+					::ftl::_dtl::call_move_and_destroy);
 
 			manager_storage.manager(
 					&other.manager_storage,
 					&manager_storage,
-					_dtl::call_move_and_destroy);
+					::ftl::_dtl::call_move_and_destroy);
 
 			temp_storage.manager(
 					&manager_storage,
 					&temp_storage,
-					_dtl::call_move_and_destroy);
+					::ftl::_dtl::call_move_and_destroy);
 
 			std::swap(call, other.call);
 		}
 
 		/// Check if function is nullary
-		operator bool() const noexcept {
-			return call != &_dtl::empty_call<R, Ps...>;
+		constexpr operator bool() const noexcept {
+			return call != &::ftl::_dtl::empty_call<R, Ps...>;
 		}
 
 	private:
-		_dtl::manager_storage_type manager_storage;
-		R (*call)(const _dtl::functor_padding&, Ps...);
+		::ftl::_dtl::manager_storage_type manager_storage;
+		R (*call)(const ::ftl::_dtl::functor_padding&, Ps...);
 
 		template<typename F, typename Allocator>
 		void initialise(F f, Allocator&& alloc) {
 
-			call = &_dtl::function_manager_inplace_specialisation<F,Allocator>
+			call = &::ftl::_dtl::function_manager_inplace_specialisation<F,Allocator>
 				::template call<R, Ps...>;
 
-			_dtl::create_manager<F,Allocator>(
+			::ftl::_dtl::create_manager<F,Allocator>(
 					manager_storage,
 					std::forward<Allocator>(alloc));
 
-			_dtl::function_manager_inplace_specialisation<F, Allocator>
+			::ftl::_dtl::function_manager_inplace_specialisation<F, Allocator>
 				::store_functor(manager_storage, std::forward<F>(f));
 		}
 
@@ -634,17 +347,19 @@ namespace ftl {
 			using Allocator = std::allocator<empty_fn_type>;
 
 			static_assert(
-				_dtl::is_inplace_allocated<empty_fn_type, Allocator>::value,
+				::ftl::_dtl::is_inplace_allocated<empty_fn_type, Allocator>::value,
 				"The empty function should benefit from small functor optimization");
 
-			_dtl::create_manager<empty_fn_type, Allocator>(
+			::ftl::_dtl::create_manager<empty_fn_type,Allocator>(
 					manager_storage,
-					Allocator());
+					Allocator()
+			);
 
-			_dtl::function_manager_inplace_specialisation<empty_fn_type, Allocator>
-				::store_functor(manager_storage, nullptr);
+			::ftl::_dtl
+				::function_manager_inplace_specialisation<empty_fn_type,Allocator>
+					::store_functor(manager_storage, nullptr);
 
-			call = &_dtl::empty_call<R, Ps...>;
+			call = &::ftl::_dtl::empty_call<R, Ps...>;
 		}
 	};
 
@@ -657,11 +372,6 @@ namespace ftl {
 	};
 
 }
-
-#ifdef __GNUC__
-#pragma GCC diagnostic pop
-#endif
-
 
 #endif
